Move parts of MDF to their own files
