+++ morphir/dataformat/src/org/finos/morphir/dataformat/mdf/Label.scala
@@ -1,0 +1,3 @@
+package org.finos.morphir.dataformat.mdf
+
+case class Label(value: String)
